refactor: 테스트 파일을 세부적인 폴더로 분리
